arm64: entry: Hook up entry trampoline to exception vectors
Hook up the entry trampoline to our exception vectors so that all
exceptions from and returns to EL0 go via the trampoline, which swizzles
the vector base register accordingly. Transitioning to and from the
kernel clobbers x30, so we use tpidrro_el0 and far_el1 as scratch
registers for native tasks.

Bug: 69856074
Change-Id: I1b03e5a5cce2be6892f5a8cd8d2857d134a74c18
Signed-off-by: Will Deacon <will.deacon@arm.com>