msm: camera: Fix eeprom driver i2c write/delay/poll issue
Eeprom i2c delay and poll cmd buf both use wait type as
their cmd type, so need to check the op_code to distinguish
the delay and poll cmd. And user space will send multiple
write cmd in a cmd buffer, so need to get all the write
cmd based on the count of header.

Change-Id: Ica18c842a7c3d02019df5b2b61f20fac239cd7c0
Signed-off-by: Depeng Shao <depengs@codeaurora.org>
Signed-off-by: Xu Han <xuhanyz@google.com>