usb: dwc3: Support switching to external vbus regulator
Opt in for external booster for direct attach devices when the
device requires VBUS current less than the current specified by
google,switch-vbus (based on bMaxPower). "google,switch-vbus"
when present and non-zero will make dwc3-msm support switching
to external vbus.

The VBUS power source is switched back to the internal PMIC whenever
the device is removed.

This patch introduces a notifier call chain that
the clients could register to know when the external regulator
needs to be used instead of the pmic's regulator.

Bug: 116661237
Bug: 67918772
Bug: 115782315
Change-Id: I78b88147cfcf1430da8579a496f760ad82b756d2
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Signed-off-by: pumahsu <pumahsu@google.com>