arm/dt: narwhal: Do not control GPIO 38 on rev-evb2-b
GPIO 38 is being used as Display Enable. However GPIO 38 is coupled
to SDC2 card detection. SDC2 CLK is affected when the state of
GPIO 38 changes. So it results in WiFi command timeout.

Do not control GPIO 38 to avoid the conflict on rev-evb2-b. It doesn't
affect the display operation. Because the Display Enable is set in
bootloader already.

Note that there is a little bit of current leakage on Display OFF. It
is about 0.04mA@4V.

Note that Display Enable will be moved to other GPIO on comming HWs

Bug: 69859012
Change-Id: I6ee190bd9c8620aad1c117dec7a1db62f7e7f6eb
Signed-off-by: Devin Kim <dojip.kim@lge.com>