staging: s2mpb04: do not clear enabled flag on regulator error
Although the regulator is disabled automatically upon an error, the
client driver still needs to call regulator_disable. That is because the
regulator framework maintains a counter of active clients. So, if a
client goes to enable it again, the regulator framework will think that
it is already enabled and will not enable it again.

Test: forced regulator failure, and then tried to power on again
Change-Id: Ief3c72299e2029e6f426d0d7258a8a22fbdf7804
Signed-off-by: Trevor Bunker <trevorbunker@google.com>