misc: mnh: prevent double ready irq disables
There is a bug in the interrupt controller where irqs stop firing after
calling disable_irq twice in a row. This patch keeps track of the state
of the irq being enabled or not to prevent this situation.

Bug: 70793461
Change-Id: I770117bc6916a3544a92a321ccbd1b715a13b12a
Signed-off-by: Trevor Bunker <trevorbunker@google.com>
(cherry picked from commit 09e35ff5076b864d52fb6699eadeea39ffd6d121)