input: atmel_u114: Fix the palm gesture not working
The palm gesture is disabled when the idle mode is set. Previously
the mode is set before the touch off and driver save the flag.
And drivers retains the setting on touch on. it results in palm
gesture not working. So reset the settings whenever touch on.

And prevent from i2c access after touch off.

Bug: 70864151
Change-Id: I01f929a28b7b926891197ed2723e9a8ef952cfd8
Signed-off-by: Devin Kim <dojip.kim@lge.com>