config: enabling CONFIG_MSM_TIMER_LEAP
Enabling CONFIG_MSM_TIMER_LEAP as per Qualcomm
recommendation for kernel crashes in scheduler.
Not sure if this will fix the kernel crash in
scheduler, but enabling this config shouldn't
cause problem because it will skip a counter when
it overflows, and it is recommended by Qualcomm.

Bug: 71724895
Test: device boots. Original issue was only reproduced
during factory testing so not tested with this CL

Change-Id: I72a126a9136e0e0cc152c63c1a74c1dde23a375f