dts: forcing all warm reboot to hard reboot
Charger and battery temps are high when doing
a warm reboot. If there is a crash then device
warm reboots, which results in higher temps
for charger and battery. The high temp will
prevent the device from being charge.

Bug: 71816520
Test: echo c > /proc/sysrq-trigger will cause the
device to crash. Before this change temps are high
and after this change temps are correct and device
charges.

Change-Id: I55e605a3133eaae15481af09cf17d667f98fc47b