leds: lm36272: Backlight off after panel off
The color level may not be consistent with the panel using RGB LED
backlight.  It's due to the charateristics of the LED response time.
It may be visible in case when display is turned off on white foreground.

Thus, in order to mitigate this problem, firstly turn off the panel and
then turn off the backlight.

Bug: 71733383
Change-Id: I20dea82df5b0a48eeff89ce64903f2b38851bb73
Signed-off-by: Devin Kim <dojip.kim@lge.com>