Limit MTP individual transfer size
Currently, an MTP transfer from Vega to a host fails because the
vfs_read call (line 883) immediately "transfers" 1mB (the value of
MTP_TX_BUFFER_INIT_SIZE). It doesn't actually happen, there is a side
effect where the USB connection is terminated until a reboot.

Based on kernel/common 4.4 the solution is to make sure the size of
xfer is 16kB.
http://cs/android/hardware/bsp/kernel/common/v4.4/drivers/usb/gadget/function/f_mtp.c?l=673

Bug: 69770261
Test: Read/write SD over MTP
Change-Id: Icc618e57d93eeabfd37d603e2a0dda66c080ab27