leds: lm36272: Turn off the backlight with DSV OFF
The panel driver controls the LM36373 IC with DSV. It turns off the IC
on blank. And then backlight driver tries to turn off the backlight.
Thus, the commands fail because the IC is already off. So control
the backlight off with DSV off.

And fix the unbalanced mutex lock.

Change-Id: Ie40a1854080ddc236eade248608840ea252f5803
Signed-off-by: Devin Kim <dojip.kim@lge.com>