leds: lm36272: Set the backlight registers whenver it needs to be updated
Remove unnecessary status checks on backlight on/off. It may causes the
registers not set properly.  For example, If trying to set the backlight
level with non-zero in case that it is already off, the try will fail and
the status, however, will be updated to ON. It will results in failure of
backlight ON with subsequential backlight level settings.

So set the backlight registers whenever it needs to be updated.

Change-Id: I2908aba8c49d971b0536959504e5b3a9a0d711d9
Signed-off-by: Devin Kim <dojip.kim@lge.com>