adsprpc: Timeout non-interruptible waits
There are some fastrpc methods that were run uninterruptible
with no timeout.  This is a bad idea since this can make the
calling process unkillable if the DSP never responds.

This CL adds timeouts to uninterruptible path.  The "kernel == 1"
path needs to remain uninterruptible since it is fatal to have the
task complete when such calls have already returned (and the state
gets cleaned up).  The "kernel == 0" path appears to need to remain
without a timeout since some invocations (eg. adsprpcd) seem to
expect to be long-running.  In comparison the "kernel == 1" calls
expect a relatively immediate response.

If the call times out, we now output a kernel message to warn that
the RPC has failed.  Also, on timeouts restart the offending
subsystem if it is one that is likely to gracefully restart.

Bug: 70381412
Bug: 65358989
Test: tango_stress_test runs, or is at least ^Cable
      "fastboot reboot bootloader" still works correctly
      adsprpcd doesn't complain of errors unless the adsp crashes
Change-Id: I9c14216de239eb8e2b23268c732ca930ae6bc422
(cherry picked from commit 0d982174196ba9c456fa8b7f36591fa24477c904)