ANDROID: qtaguid: Fix the UAF probelm with tag_ref_tree
When multiple threads is trying to tag/delete the same socket at the
same time, there is a chance the tag_ref_entry of the target socket to
be null before the uid_tag_data entry is freed. It is caused by the
ctrl_cmd_tag function where it doesn't correctly grab the spinlocks
when tagging a socket.

CVE-2017-17770

Change-Id: I5307e7da1eb911f0589f77d5832bb45dac1bccaf
Signed-off-by: Chenbo Feng <fengc@google.com>
Bug: 65853158