power_supply: fix POWER_SUPPLY_PROP_OTG_FASTROLESWAP type
The enum power_supply_property is sorted by type, and the function
power_supply_show_property relies on this order to determine which
field of the power_supply_propval union is valid.

POWER_SUPPLY_PROP_OTG_FASTROLESWAP is currently grouped together with
string values, which leads to an occasional kernel panic when reading
the otg_fastroleswap sysfs entry, which is actually an integer.

Bug: 72648767
Change-Id: Ic0452758ce6e7052e8ecb89ab72299d8212ad894
Fixes: b225429ece2c ("smb-lib: Add property to enable/disable otg threshold")
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>