misc: sop716: Add new commands and enhance fw update
Add the command revision info for every command. The commands are
extended with firmware update. If fw doesn't support for specific
command, the command will be ignored simply.

There was no way to know the current mode during fw update. Now
driver can read the mode of movement and restore watch mode with
using it.

Change-Id: I9ec6b1f63da89e8a2eebe253e19eaf97df7e848a
Signed-off-by: Devin Kim <dojip.kim@lge.com>