Remove id from console-ramoops file.
BootReceiver.java checks for console ramoops logs without an id attached
to the name. In order to generate crash reports for device telemetry we
need to drop the id from the crash report. We have only ever seen id:0
on these crash logs as well so we are not losing data via this change.

Bug: 73006871
Test: Device boots, kernel crash creates a console-ramoops file that
does not have an ID appended to it.

Change-Id: Id8d5d969ecee29e48e7dbdc89605809bba749d45
(cherry picked from commit ce666df90425998ba64c42bf42a05f340ebb002c)