clk: Add the CLK_CHILD_NO_RATE_PROP flag
The voter clock mechanism relies on the rate request
propagation to the parent clock to be one-way and not
have the clock framework propagate the rate across all
its children. Add the CLK_CHILD_NO_RATE_PROP flag for
these votable clocks to use in order to avoid the top-down
propagation of any scaling requests.

Porting over to floral since we're seeing two-way voting
propagation for the l3_clk. clk_set_rate() is seen to
have top-down propagation after a child requests its
parent change it frequency. We had this for 4.9 but
have not forward ported it to 4.14.

Test: boot
Bug: 129537495
Bug: 132350331
Change-Id: I757321ae43079a6f4fd05ebdc5cf3d4f388a3fab
Signed-off-by: Martin Liu <liumartin@google.com>