arm/dt: narwhal: Lower the movement EOL limit to 3300mV
The voltage drops happen during device power off in low battery.
Thus the EOL limit reaches in this case and movement may shutdown
unexpectedly. So lower the limit to avoid it. And the logic for checking
the limit will be improved in FW level in order to avoid unexpected shutdown
by sudden voltage drop as well.

Bug: 73000248
Change-Id: Ic94905444dfe64d4da4db2884ec9f18a24fb765a
signed-off-by: Devin Kim <dojip.kim@lge.com>