arm/dt: narwhal: Disable vdd-always-on on eMMC
Previously kept vdd being always on due to the write protection (aka WP).
If the vdd is allowed to be off at the runtime, emmc won't retain the WP
after that.

There is a side effect, which is not working on emmc hard reset, with
the vdd-always-on. It causes hang during emmc hard resset and results in
a watchdog crash.

Narwhanl doesn't need the eMMC write protection feature. RPMB will be used
insead. So allow vdd to be off again.

Change-Id: I29eee8efe98307188776087d44878cecb4b6963e
Signed-off-by: duhan.kim <duhan.kim@lge.com>
Signed-off-by: Devin Kim <dojip.kim@lge.com>