Disable CPU sets on Narwhal
The use of CPU sets can cause SurfaceFlinger to be de-scheduled in favor
of lower priority tasks.

When CPU sets are enabled SurfaceFlinger is added to the
'system-background' CPU set which is limited to CPU 0 once boot has
finished. If another high-priority thread (such as the RenderThread) is
scheduled on CPU 0 SurfaceFlinger may be delayed, leading to jank.

Bug: 68943050
Change-Id: I884fbe3ad91cf55747af896a9243224f7dedebd1
Signed-off-by: Matt Bouyack <mbouyack@google.com>