msm_serial_hs: Add defensive checks due to sudden port close
BACKPORT based on QC commit ab1ab13f858e6ffaff33defd3b29781144e046db

In certain scenarios, it is possible for a close to be called on the tty
port while the rx thread is still processing. In certain scenarios a
race between the 2 threads could lead to unexpected results such as making
BAM API calls after the pipes are disconnected or the thread accessing the
framework structures after they're freed.
Add defensive checks to:
- Check the state of the rx pipes and tty before running the rx thread.
- When closing the port, flush the rx thread and wait till there
  are no pending buffers before forcing a close. Currently the shutdown
  waits on the flush state of the rx pipe to be in SHUTDOWN, but this
  only happens after the rx thread runs, kicks the runtime pm which puts
  it in the SHUTDOWN state. But since the port close calls the runtime pm
  suspend callback, just wait on the thread to finish the buffer.

Change-Id: I561865c0282b34fa758d965b3e6eb3eb8b9e6ee4
Signed-off-by: Devin Kim <dojip.kim@lge.com>