arm64: dts: Disable null DSI packet insertion on B1
Null packets are a feature of DSI that keep the data lanes in high
speed mode.  This is necessary for some display geometries, but
not for B1.  In fact, B1 experiences display corruption due to
insufficient clock toggling on the DSI lanes when null packets are
inserted.

This patch removes the device tree property enabling null packet
insertion on B1.  It continues to be enabled on C1 as before.

Bug: 73170320
Change-Id: I6458fb679ff60ba72ffa4eac96da4a150fa4226c
Signed-off-by: Sean Callanan <spyffe@google.com>