Revert 'msm: mdss: fix fence time out issue in MDP3 driver'
The change (82321a1bebb1) significantly increased the likelihood of
jank. Specifically it causes us to block in hwc_sync while waiting for
the previous frame to be DMA'd to the panel.

Note that a pure revert of the change introduces a new bug in which the
devices randomly reboots when entering suspend, so this change includes
the fix for that as well.

With the revert we would no longer signal MDP_NOTIFY_FRAME_CTX_DONE
which could cause us to block for 30 seconds in mdp3_ctrl_off waiting on
'kickoff_wait_q'.

Instead we retain the code which signals kickoff_wait_q and remove the
check which makes it conditional on 'wait_for_kickoff'.

Bug: 68943050
Bug: 74620493
Bug: 77597682
Change-Id: I83f855cfd3d6f67579dcf127ddcd402d464d1d7d
Signed-off-by: Matthew Bouyack <mbouyack@google.com>
(cherry picked from commit 6f8837511d54666905c29ba8eb9420a7540e4de7)