msm: camera: reqmgr: Update frame sync logic for dual camera
Currently kernel frame sync expects that SOF's from one sensor
will always lead with respect to the other. This change removes
this dependency, but at the same time it is tightly bound with
HW frame sync. Also if a packet from a node is not available at
a given SOF for a particular link we will skip that frame and as
well as the corresponding frame in the sync link to ensure we
output matching frames on both links.

Bug: 118363163
Test: GCA, CTS

Change-Id: I5aad699302c163d75878b139ab8f079f8370886e
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>