msm: camera: sensor: Fix sensor driver entering infinite loop
Needed when the writes in array has continuous addresses.
In i2c write sequence, if there are continuous
addresses, we are using batch mode to do I2C operation
by clubbing them.
But if the length of batch goes beyond ‘payload_size’
we are returning error. Instead, we need to do multiple
batch operations with max of payload-size each time
without returning failure.

Bug: 118363163
Test: GCA, CTS

Change-Id: I8eb55102becf372f358b2ac2a2e80a563a34e577
Signed-off-by: Vishalsingh Hajeri <vhajeri@codeaurora.org>