usb: tcpm: Correct the calculation of PD current
The calculation of PD current (milliampere) after selecting the PDO is
incorrect. If max_mw is not divisible by max_mv, the calculated value
of ampere will be rounded down. The discarded current value may be up
to almost one ampere. Fix this by converting the max_mw to microwatt
first.

Bug: 77361040

Change-Id: I71a50e2dd31a4ac2ab95775f14d49a33aa277ef4
Signed-off-by: Kyle Tso <kyletso@google.com>