power: supply: max1720x: rework config management
Simplified config management to just by registers instead of individual
naming.
Doing a full read + copy + update + writeback is != takes ~200ms,
and this time fill be mostly constant.
Doing a per register read + compare for ~60 register takes 400ms and
strongly depends on the number of registers to check.
Register values as expected to be less so doing a per register check.

Bug: 75287390
Change-Id: Ic02d4de9ac02e6c345a9ea9900cb08050541eb2b
Signed-off-by: Thierry Strudel <tstrudel@google.com>