leds: lm36272: Terminate the bl work immediately if new bl update
bl work is protected with a mutex. Thus if it's runing on new bl update,
upperlayer will wait for the bl_work done. Unless Android framework
gets a response in a determined time, it will force to turn off the
display. To avoid this, terminate the bl work immediately if there is
a new bl update.

Bug: 72784365
Change-Id: I50c8c7bea24496e957d8daf8c6ce2985a4d0b797
Signed-off-by: Devin Kim <dojip.kim@lge.com>