power: supply: max1720x: rework config management
Simplified config management to just by registers instead of individual
naming.
Doing a full read + copy + update + writeback is != takes ~200ms,
and this time fill be mostly constant.
Doing a per register read + compare for ~60 register takes 400ms and
strongly depends on the number of registers to check.
Register values as expected to be less so doing a per register check.
Add an i2c_driver remove function to release the iio_channel on exit
and not when done.

Bug: 75287390
Change-Id: Id127f959b0ee72e5c0fc53dbada74f3a181f3d5b
Signed-off-by: Thierry Strudel <tstrudel@google.com>