Makefile: append BUILD_NUMBER to version string when defined
In current implementation the Linux version string that is attached to
the kernel linux image as well as the all kernel modules includes the
sha1 of the main git project for the kernel.

This Commit adds the BUILD_NUMBER env variable with an 'ab' prefix to
the Linux version string. This maps to both the Linux Image and all
kernel modules version strings.

Bug: 78243318
Test: Build the kernel and test the version string

Change-Id: I4432f3463e0e7ca55eb1f4812562f56a002e2a31
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>
(cherry picked from commit 527b0ca58da9366337e9692bf3e00c4d745b3849)