build.config.kasan: Set CONFIG_FRAME_WARN to 0
CONFIG_FRAME_WARN needs to be set to 0 instead of being disabled.
Otherwise, if it's just disabled, kconfig will set it back to the default
which is 2048 for 64 bit targets.

Previously, the default was 0 if CONFIG_KASAN is set, but the
stable commit ef3af3465ab4 ("kasan: rework Kconfig settings") has
removed that default for the CONFIG_KASAN case.

Change-Id: I4f9cd5a7a3d8400fe5a48e83bb0dc8db4f9dbdf9
Signed-off-by: Daniel Mentz <danielmentz@google.com>