cpufreq: times: concurrent_policy_time by pid
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each pid spent running
on group of cpu's with the same policy with 1
to num_related_cpus online and not idle.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 2 cores that are a part of the
same policy are online and active for 50ms and
the pid is running on one of the cores, the pid
should be blamed for 1/2 of the energy that is
used everytime one or more cpus in the same policy
is active. Using other proc files, the pid can
also be blamed for the cost of keeping the entire
cpu unsuspended and the cost of the individual
cores it is using.

This patch is based on commit
7afc0cf00242a00dcde448aa83fca5d724826748 in the
android-msm-wahoo-4.4 branch by Marissa Wall.

Test: cat /proc/<PID>/concurrent_policy_time
Bug: 68399339
Bug: 74242328
Change-Id: I12334d53cff6006f42df1b6ae744871d596fd0b1
Signed-off-by: Connor O'Brien <connoro@google.com>