typec: tcpm: fix compiler warning about stupid things
gcc thinks it is too smart and gives off a "you might be using this
variable before it is initialized" warning in tcpm_pd_build_request(),
because it can not follow the logic through the tcpm_pd_select_pdo()
call.

So just make gcc quiet by initializing things to 0, to prevent the
myriad of people complaining that we now have a build warning.

Bug: 112400445
Change-Id: I15b2f8472dc93a6ceb091df5a987f109f050d512
Cc: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Cc: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Cc: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: pumahsu <pumahsu@google.com>