input: touchscreen: stm Use threaded irq
Currently, when touch interrupt is received, work is scheduled to read
the data on the i2c bus. This could lead to potentially long i2c reads
that do not get fixed with requests to pm_qos (next patch). Instead of
scheduling work, the processing can be done right away in the bottom
half of the interrupt handler, but with other interrupts enabled (so
latency is not affected). This also allows usage of pm_qos, and
simplifies the driver code a bit.

After the full set of patches, the worse case scenario is:
1.3 ms delay between interrupt and the start of the first i2c read
2.4 ms i2c read duration for the 8 bytes of event data

Test: write an app that displays full-screen black image; wait a bit
(10-20 seconds), however, force the screen to be on while on usb. Then,
run the following:
./external/chromium-trace/systrace.py --atrace-categories=core_services,i2c,irq,hal,input,wm,am,view,power,idle,sched
and analyze the trace.

Bug: 79422357
Change-Id: Ie7f70cd71523bb51e2d700c911a2cda188fd9c0d
Signed-Off-By: Siarhei Vishniakou <svv@google.com>
Signed-off-by: Steve Pfetsch <spfetsch@google.com>