input: touchscreen: stm Require low latency
Currently, CPU can enter deep sleep while waiting for data on the i2c
bus. That's bad, because it will then take a while to wake up after
receiving the data. This, in turn, may mean a) touch event handling time
will be longer b) the touch handler may not complete before the next
interrupt comes in.
After discussing this issue with Qualcomm, the solution is to use pm_qos
to demand prevent CPU from entering deep sleep.

The remaining delays that we are experiencing now seem to be:
1) After the CPU has been sleeping for a bit, it takes some time for it
to wake up after receiving the interrupt.
2) There *may* be some hardware that also needs to wake up to service
i2c bus, which is why we are seeing that it takes ~ 2.5 ms to transfer 8
bytes of data on the bus when the very first read happens.
These delays currently appear to be acceptable given the touch report
rate of 100 Hz.

Bug: 79422357
Test: systrace (see parent commit)
Change-Id: Iaa28be45c883e052655c62937e9b7c17b1ecbeaf
Signed-Off-By: Siarhei Vishniakou <svv@google.com>