arm64/config: b1c1: Enable kprobes
This feature is needed for BCC to work on bluecross devices.
For more details, see: go/bcc-android

kprobes has no overhead unless its in use and is a common feature
enabled in production kernels.

It is possible to use it only from kernel modules or kernel code running
in a privileged environment.

Bug: 77809405

Change-Id: I0a13e0488ab6ff5387d72915ee8ac6bb0f032604
Signed-off-by: Joel Fernandes <joelaf@google.com>