input: touch: tbn: suppress warning on QMI payload
We are only using QMI message header for the switching protocol. We
added payload field in QMI response because kernel QMI decoder will
crash if response does not contain a payload. We had a slight IDL
mismatch as we don't have access to common.idl which defines the
response struct, but since we don't care about the payload field, it
should be safe to remove the decoding warning on the payload.

Bug: 79252467
Test: verify warning is gone and everything works fine
Change-Id: I29bbf42a8a7ce578acc1b113b37d379f43a80d93
Signed-off-by: Max Shi <meixuanshi@google.com>