input: touchscreen: stm: add MSC_TIMESTAMP
Acquire timestamp during hard interrupt portion of the touch event and
report it as MSC_TIMESTAMP.
MSC_TIMESTAMP is in microseconds. Given that it is 32 bits in size, it
will overflow after a little over an hour.
Acquire the timestamp in nanoseconds, to be used later by v4l2, but
report here in microseconds as per the input spec.
Together with input_event timing information, the userspace can still
reconstruct the original time of event up to a microsecond.

Consumers of the /dev/v4l2-touchX (coming in a future commit) will have
access to the full timestamp in nanoseconds.

Example stream of light touch after patch:
[.990824] /dev/input/event2: EV_ABS       ABS_MT_TRACKING_ID   000000f9
[.990824] /dev/input/event2: EV_ABS       ABS_MT_POSITION_X    000001ee
[.990824] /dev/input/event2: EV_ABS       ABS_MT_POSITION_Y    0000034c
[.990824] /dev/input/event2: EV_MSC       MSC_TIMESTAMP        139bfc07
[.990824] /dev/input/event2: EV_SYN       SYN_REPORT           00000000
[.997656] /dev/input/event2: EV_ABS       ABS_MT_TRACKING_ID   ffffffff
[.997656] /dev/input/event2: EV_MSC       MSC_TIMESTAMP        139c190b
[.997656] /dev/input/event2: EV_SYN       SYN_REPORT           00000000

Bug: 62940136
Test: getevent -lt
Change-Id: I0ae3553b2f3a8e7e13cccb3add6bd9cf250e75bc
Signed-Off-By: Siarhei Vishniakou <svv@google.com>