Add timeout and panic when __scm_call_armv8_64 takes too long.
Device can stuck in the infinite loop in __scm_call_armv8_64 if TZ
subsystem stays in SCM_INTERRUPTED state. To get logs while the device
is in freeze state, added a timeout in __scm_call_armv8_64 and triggered
panic if SCM call takes too long.

This should be reverted when root cause in TZ is identified.

Bug:74358862
Test: Ran the kernel on device for >4hr without a crash.
Change-Id: I9ee3c592fca23209b72004dd95cabc51e398eb3b
Signed-off-by: Siyuan Zhou <siyuanzhou@google.com>