msm: camera: eeprom: Correct free call for power settings
Currently memory for the power settings is allocated whenever
we receive an init packet in packet parser but that memory is
freed only during device close leading to repeated allocations.
This change ensures that we free the memory as soon as we power
down the device avoiding repeated allocations of memory.

Bug: 118363163
Test: GCA, CTS

Change-Id: I7fac6498c9403e8ae913dfc4f20f9aa8f0bdb999
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>