tcpm: Prevent orientation loss during PR_SWAP
When a port is a DRP and the toggling is run
by hardware, the hardware controller might
start to toggle as soon as vbus falls below the
disconnect threshold. Fix this by asserting
the Rp/Rd before beginning the role_swap sequence.

In addition, call the in_pr_swap callbacks for
the lower layers to implement board specific
operations

Bug: 116661237
Bug: 80001326
Bug: 140367237
Change-Id: Ie87739b921135c2e9d83336cc03e59540379e861
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Signed-off-by: Kyle Tso <kyletso@google.com>