soc: qcom: smp2p: Introduce pending state for virtual irq
If a smp2p change occurs while a virtual interrupt is disabled, smp2p
should be able to resend that interrupt on enablement.

This functionality requires the CONFIG_HARDIRQS_SW_RESEND to be enabled
to reschedule the interrupts. To ensure the mask and unmask functions
are called during enabled and disable, set the flag to disable lazy
IRQ state handling (IRQ_DISABLE_UNLAZY).

Change-Id: I75c662c61a9705c05278d238ca51f7a2e74806d8
Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Qilong Wang <quic_qilowang@quicinc.com>