wlan: fix possible OOB read in SIOCSIWRATE WEXT ioctl handler
The values used to dictate the end of the for loop is great than the
size of array supp_rates, this will cause an OOB read when loop
through supp_rates. So need modify the size of array supp_rates.
There's also a functional issue in that the second call to
sme_cfg_get_str() overwrites the lower values of the first call,
thus not ever allowing the lower channel rates of A to ever be
valid. So need update the read buffer address for the second
sme_cfg_get_str().

Change-Id: I772042e1caa14c49c54eac2f04e854cb4c9ac9de
CRs-Fixed: 2247498