drm/msm: track msm_gem_obj in active_list with a boolean
The active_list is maintained to keep track of the unmapped
buffers during the detach of a context-bank, so that it can
be re-mapped during the re-attach process. Add a boolean to
track if the msm_obj is part of active_list to avoid adding
it twice. In addition, set the aspace domain_attached flag
before the re-mapping of buffers during the re-attach process.
This is required to re-map the buffers for which the domains
were removed during detach of the context-bank.

Bug: 79994236

Change-Id: Ic4daa020dbed2faaffd609ceb369a2fbd36af4cf
Signed-off-by: Veera Sundaram Sankaran <veeras@codeaurora.org>
Signed-off-by: Roopesh Rajashekharaiah Nataraja <roopeshr@codeaurora.org>