b1c1_defconfig: disable CONFIG_EDAC_KRYO3XX_ARM64_PANIC_ON_CE.
Single-bit corruption would cause kernel panic with this config.
Since occasional single-bit errors are expected in normal operation, and
there is enough redundancy in the encoding that HW can recover the content
with error correction code automatically, the config should be disabled
to avoid device crashing.

The config was disabled on 2017 devices.

Bug: 80369630
Test: verified device boot up and /proc/config.gz shows the config is not set.
Change-Id: I9f5156dd281ab9182f8c3cf22de5cf058a7ca5cd
Signed-off-by: Siyuan Zhou <siyuanzhou@google.com>