display: not to disable the regulator, L6 in recovery mode
We found out that this regulator would be turned on and off a couple of times
as the device boots into recovery mode.
In such case, touch driver may disable its own IRQ and end up a dysfunction.

Hence, L6 could not be disabled for preventing touch failure in recovery mode.

Change-Id: I62ba2a79ea8565bd60145a619bc462aec4dbfccd