pd_engine: OTG co-ex wireless charging
Limit the mode of operation to USB2.0 and use TPS
to source vbus when wireless charging is enabled.

Wireless supply's POWER_SUPPLY_PROP_ONLINE prop is
used to infer the status of wireless charging.

EXTCON_PROP_USB_SS is used to limit the mode of
operation of the controller.

(Port of ce094d68a8dea8f30a256425685ecc16d0314da1)

Bug: 77536221
Bug: 140367237
Change-Id: I37c3677547859478698165c42845b83890ca9491
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Signed-off-by: rickyniu <rickyniu@google.com>