Revert "Revert "misc: mnh-pwr: Update link status first on emergency shutdown""
This reverts commit 3be71e72345e8aa708b0eeb93fa319c9519f1e21.

On an emergency shutdown event, link may be in a bad state and may
cause kernel BUG on our platform:
  Unhandled fault: synchronous external abort

In such case, update PCIe link status before disabling PCIe irq,
in order to avoid any read/write to PCIe configuration registers.

In the normal case, the original sequence is not affected.

Bug: 76434914
Change-Id: Ib6f9f2a22e3076ebffe643d928c8ddbb34199b3f
Signed-off-by: Cheng Gu <gucheng@google.com>