misc: mnh-pwr: Update link status first on emergency shutdown
On an emergency shutdown event, link may be in a bad state and may
cause kernel BUG on our platform:
  Unhandled fault: synchronous external abort

In such case, update PCIe link status before disabling PCIe irq,
in order to avoid any read/write to PCIe configuration registers.

In the normal case, the original sequence is not affected.

Bug: 76434914
Change-Id: I0ef078c747660cbb31e57142cac12bfb6a2b3162
Signed-off-by: Cheng Gu <gucheng@google.com>