easelcomm: defend against concurrent DMA requests for the same message
Add concurrency protection for scatter-gather list create and destroy
calls.

Return -EBUSY on a duplicate attempt to send or receive DMA for a
message that already has a scatter-gather list setup; do not
overwrite the pointer to the existing scatter-gather list, and remove
the ability to free a scatter-gather list in use by another thread.

Bug: 69808833
Change-Id: I8e9366f290817c0065b1e3c33453d96057be3172
Test: Bug 69808833 POC
Test: EzlshTest
Test: manual: EaselCamera
Test: manual: GCA-eng
Signed-off-by: Todd Poynor <toddpoynor@google.com>