power/BCL/Easel: Added throttling Easel to BCL.
Added binding of mnh cooling device to SOC/Vbat/ibat thermal zone.

Bug: 80547368
Test: Confirmed associated cooling device cur_state at thermal trip point.
For example: mnh's cur_state should be 0 when it is not throttled and 3
when it is throttled.  Thermal trip point of SOC = 10% is used.
Preliminary measurement: 500-600mA current saving when Easel is
throttled.  Setting state to 3 is not required at the time of this CL.
This changelist merely puts in the framework for enabling Easel
throttling if needed.
Change-Id: Iff1815b87304a3106d13a5b75991731bc3782f38
Signed-off-by: George Lee <geolee@google.com>