input: touchscreen: stm: support pressure / size / btn_touch
Currently, pressure and size is not being reported by the driver. Also,
BTN_TOUCH is not currently properly reported by the driver. Recent patch
to enable the pressure reporting caused a touch regression, where a
device was unresponsive to touch. This is likely because the device was
getting into a strange state where it would think that a pointer is
always present. With this patch, have not been able to reproduce that
issue.

The tool type, BTN_TOOL_FINGER, is not currently being reported. With
this patch, that code will be more correct, in the case we do
decide to enable this reporting. Given that this report is optional per
https://source.android.com/devices/input/touch-devices, the touch should
still operate without this information. It is also not clear whether the
firmware actually supports tool types other than finger.

Bug: 74216515
Change-Id: I90ef818a710c370f0d0905df4dee9dff556b0b4c
Signed-Off-By: Siarhei Vishniakou <svv@google.com>