max1720x: use FullCapNom to report full capacity
FullCapNom is the battery capacity without any temperature or empty
compensation while FullCapRep includes both. This implies FullCapRep
will vary more over time including based on temperature of charge and
complete or incomplete charging as compared to FullCapNom. Accuracy of
FullCapNom is expected to be on the order of the current accuracy
measurement of the fuel gauge (e.g. +/-1%).

Bug: 110373809
Change-Id: I9e640a960865f354262640f98d31f1c32d5b83ff
Signed-off-by: Thierry Strudel <tstrudel@google.com>