pd_engine: refresh power_supply state
When PROP_IN_PR_SWAP is set, smb-lib disables
the usb attach/detach interrupts to maintain
the orientation. This implies that the PROP_PRESENT
of vbus is not refreshed either. Re-query all the
usb psy properties to refresh the state when
PROP_IN_PR_SWAP is set to false again.

Bug: 116661237
Bug: 80001326
Change-Id: Ia6606620b91fdbe9e885f3cc3e533be908276b90
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>