s2mpb04: read interrupt status regiter beforing power on
Emergency shutdown status (e.g., from watchdog, from thermal trip) is
sticky even across reboots as long as VBAT is maintained. If any of
these status bits are set, the s2mpb04 will not deassert RESETB. This
change adds a read of the interrupt status registers before asserting
power on to clear any pending status bits.

In addition, in at least one case, we saw that the device had the
watchdog expire status bit set. The watchdog is unused and we are still
investigating how it ever got set and triggered. Until then, we are
explicitly disabling the watchdog after power on.

Bug: 110270380
Change-Id: I84d87b18dc480fe2b7c9de9e7e7c00dcabac6986
Signed-off-by: Trevor Bunker <trevorbunker@google.com>