power: p9221_charger: Add FOD settings for EPP mode
Add separate FOD settings for EPP mode and add to status
output for verification.

Allow number of FOD parameters to be specified in the device
tree.  Not all of the FOD parameters need to be loaded,
the ones which are not loaded will be left as default.

Bug: 73298980
Bug: 80388497
Change-Id: Ieada3587ab7bf2c4e8eb8eaab86afe64d2ac5e20
Signed-off-by: Patrick Tjin <pattjin@google.com>
(cherry picked from commit f8bb4cea2909a7d2bab7e481e678459ccb9b4693)