mnh-sm: Report wakeup event in mnh-sm
On AP system suspend, PCIe endpoint will be reset.  In some rare
background tasks, we want to make sure Easel stays awake during
the task and temporarily prevent AP from suspend.

Example of such tasks are stopping service in background and
updating firmware.

In this patch, a wakelock is held with a timeout when Easel is
active.  Driver automatically releases wakelock after timeout
value, or releases wakelock whenever Easel is not active.

Test: GCA
Test: cat /sys/kernel/debug/wakeup_sources
Bug: 110381808
Change-Id: I73b236ff6cee95bdf9f358e01c2c4555ec29e14c
Signed-off-by: Cheng Gu <gucheng@google.com>