power: p9221_charger: Write FOD at DC-IN
Writing the FOD too soon after VRECT is not effective as
the parameters are overwritten before DC-IN.  This patch
moves the writing of FOD until right after DC-IN is detected.

The FOD parameters are not used until RPP packets are sent
which will be some time after DC-IN.

Also add retry mechanism for validating written FOD and clean
up version output.

Bug: 110245589
Change-Id: Ic21349f92f30f16b1610fcc02a392ff1fe5ae473
Signed-off-by: Patrick Tjin <pattjin@google.com>