power: p9221_charger: reset charge status when coming online
If the user takes the device off of the charger and puts it back on
but the charge status had not changed, the driver would not send an
updated CSP packet.  Reset the last status when coming online and
only store the last status if we actually sent it.

Also, fix the R5 detection.  R7 chipsets still return CID=5.  Default
to R5 since that is the production version.

Bug: 80435107
Change-Id: I3ccca94ea424bc1a079bd2e2d0b580307dac5b90
Signed-off-by: Patrick Tjin <pattjin@google.com>