drm/msm: dsi-staging: allow LP brightness compensation
sw43408 panel has a register that allows brightness compensation during
Low Power mode. However this register varies for every panel so we need
to read the register before updating it.

Provide a mechanism to allow an adder to be added to the register to
compensate for entering low power mode and boost brightness.

Bug: 79539541
Change-Id: Iadf242c719f0709e6b2ba3d24bd7e7e6a6b92610
Signed-off-by: Adrian Salido <salidoa@google.com>