mm: add ION heap stat to meminfo
I'm not sure we really need because generally, vmstat is better than
meminfo so it might be enough with vmstat but unfortunately,
android doesn't track vmstat at this moment and lmkd uses meminfo
rather than vmstat so let's add ION heap allocation stat to meminfo.

With that, lmkd could print snapshot of memory included ION easily
when it kills the app.
It would be helpful to debug LMKD killing problems.

Bug: 110330255
Change-Id: I1c95d0e11ad416af841c9409cedc47ec615ab666
Signed-off-by: Minchan Kim <minchan@google.com>