smb-lib: Put back the hvdcp_timeout
After HVDCP is unsupported, PD state machine no longer relies on the
hvdcp_timeout to be allowed to start. This will cause PD state machine
to run while APSD is in progress. If PD port partner support USB host
communication, the USB enumeration might interfere with the APSD.

To tackle this, put the HVDCP_TIMEOUT_VOTER back to the decision of the
starting of PD state machine. And force it to vote for PD allowance
after DCP is detected.

Bug: 111241955
Change-Id: I501283ec75c14506b6cd49114ff1c7117321b7c2
Signed-off-by: Kyle Tso <kyletso@google.com>