msm: camera: isp: Fix race condition in tasklet stop
This change sets the atomic variable tasklet_active to 0 in
stop before calling flush. This also adds call to tasklet_kill
before tasklet_disable in stop. This will fix a race condition
between tasklet stop/enqueue. And it will ensure the tasklet
is correctly halted and will not be raised repeatedly.
Bug: 110695237

Change-Id: I888f7cff29e8eb0a6bfad26f081ff67e83c1d012
Signed-off-by: Harsh Shah <harshs@codeaurora.org>