pinctrl: qcom: Add support for shared SPI config for GPIOs
For all the shared peripheral interrupts, apart from the type
configuration done in the interrupt controller, the type also
to be programmed accordingly through the shared spi config
registers. These were statically done for all SPIs.

But for all the GPIO interrupts connected to the GIC, their
type need to be set dynamically, as the usage of particular
GPIO would vary. So set the shared spi type configuration
for the GPIOs accordingly when set type is called for it.

Change-Id: I1fd1b37bdcebcbfec641e4db04cb29f9e087d426
Signed-off-by: Srinivas Rao L <lsrao@codeaurora.org>