input: touchscreen: sec_ts/stm: Early disable
Disable touch earlier. Currently, touch is disabled after the display
has turned off. That means that the user may think the screen is off,
and then (during a phone call), accidentally trigger some UI action with
the ear. If the touch is disabled before screen is turned off, then it
will be clear to the user that the screen is still active, and to be
more careful.

Test: add logging to indicate MSM_DRM_EARLY_EVENT_BLANK and
MSM_DRM_EVENT_BLANK. Use power button to toggle screen on and off. Check
that the timing has improved with the patch, as the touch is getting
disabled during MSM_DRM_EARLY_EVENT_BLANK now.
Test: Built an app that uses PROXIMITY_SCREEN_OFF_WAKE_LOCK, similar to
what the dialer uses. Manual testing with the app. Also, confirmed with
systrace that when touchscreen is being pressed continuously, and the
display turns off, then the time difference between the display off
command and the last touch irq decreases.
Test: ensured double-tap-to-wake still works
Bug: 112195471
Signed-Off-By: Siarhei Vishniakou <svv@google.com>

Change-Id: Ic7db5bd45c4aacf34d301c2efadcaa53ff43d2ff